feat(shared-rasters): migrate build_border_dem + build_vpu_landmask + build_vrt to library mode
Second production-flow migration cluster. Moves the three remaining
post-prep raster steps (border-DEM fill, per-VPU HRU landmask, CONUS VRT
assembly) into builder modules under gfv2_params.shared_rasters/ so the
unified orchestrator can run them in DAG order.

After this commit, 7 of the 9 PLANNED_STEPS entries are registered:
  merge_rpu_by_vpu, compute_slope_aspect, build_border_dem,
  build_vpu_landmask, compute_dem_derivatives (opt-in),
  merge_rpu_by_vpu_twi, build_vrt
Only build_derived_rasters and build_lulc_rasters remain (cluster 2e).

Key design notes:
- build_border_dem and build_vrt are CONUS-once (do not iterate ctx.vpus).
  They discover per-VPU sources by globbing nhd_merged/ rather than
  consuming a VPU list. ctx.vpus is set to [] in their CLI shells to
  reflect this; the builders ignore the field.
- build_vpu_landmask is per-VPU. Its template_raster + output_raster
  paths in shared_rasters.yml use the {vpu} placeholder, which the
  builder substitutes inside its iteration loop.
- The orchestrator now treats {vpu} as a deferred placeholder
  (_DEFERRED_PLACEHOLDERS = {"vpu"}). Unresolved {vpu} at config-load
  time is allowed; any other unresolved placeholder still raises with
  a clear message listing both available-now and deferred-to-builders
  sets. Mirrors how the depstor orchestrator defers {fabric}.
- build_vrt returns its 5 produced VRTs in ctx.paths (elevation_vrt,
  slope_vrt, aspect_vrt, fdr_vrt, twi_vrt) so future builders or
  downstream tooling can look them up by name.
- build_border_dem returns {border_elevation, border_slope, border_aspect}
  for the same reason.
- build_vpu_landmask returns {} — per-VPU outputs are conventionally
  pathed and downstream builders re-template rather than central lookup
  (consistent with merge_rpu_by_vpu / compute_slope_aspect).

scripts/build_vpu_landmask.py re-exports build_vpu_landmask() (the public
helper used by tests/test_build_depstor_landmask.py and friends) from the
new library module, so existing test imports keep working without
touching the test files.

Verified end-to-end:
- Unified orchestrator successfully ran the build_vrt step against the
  real data_root: rebuilt all 5 CONUS VRTs (elevation/slope/aspect with
  18 NHDPlus VPUs + 1 Copernicus fill source; fdr/twi with 18 NHDPlus
  only — no border fill, as expected) in 5s.
- Legacy build_vrt.py CLI rebuilds the same 5 VRTs (compatibility shim
  works).
- Legacy build_vpu_landmask.py CLI correctly detects existing land mask
  for VPU01 and skips.
- All 7 orchestrator test invariants still pass; only 2 of the 9
  PLANNED_STEPS remain pending (build_derived_rasters, build_lulc_rasters).

Step 2d of 6 on the consolidation branch.
