+++ scripts/build_border_dem.py
@@ -1,220 +1,20 @@
 """Build Copernicus GLO-30 elevation fill for border HRUs (Canada/Mexico).
 
-Downloads Copernicus 30m tiles covering border zones, mosaics them,
-reprojects to EPSG:5070 at 30m, then builds a composite elevation surface
-by overlaying NHDPlus VPU tiles on top of Copernicus (NHDPlus takes priority
-in the overlap zone via GDAL VRT last-source-wins ordering). Slope and aspect
-are computed via RichDEM on this composite, then masked to retain only pixels
-in the fill zone (where Copernicus has data but NHDPlus does not).
+Thin CLI shell over the ``gfv2_params.shared_rasters.build_border_dem``
+library entrypoint. Preserves the original argparse interface so existing
+sbatch jobs keep working unchanged.
 
-Output tiles are placed in work/nhd_merged/copernicus_fill/ where
-build_vrt.py lists them before NHDPlus tiles in the VRT (GDAL
-last-source-wins means NHDPlus takes priority).
-
-Dependency: must run AFTER compute_slope_aspect.py (per-VPU), because it
-needs the NHDPlus _fixed_ elevation tiles for the composite.
+For unified DAG-style invocation across all shared rasters, prefer:
+    python scripts/build_shared_rasters.py --config configs/shared_rasters.yml
 """
 
 import argparse
-import time
 from pathlib import Path
 
-import numpy as np
-import richdem as rd
-from osgeo import gdal
-
 from gfv2_params.config import load_base_config
-from gfv2_params.download.copernicus_dem import download_tiles, tiles_for_bbox
 from gfv2_params.log import configure_logging
-
-# Border bounding boxes in EPSG:4326 (south, north, west, east).
-# Deliberately generous — extra ocean tiles are skipped (404) and
-# NHDPlus takes priority in overlapping areas because build_vrt.py lists
-# Copernicus fill tiles before NHDPlus tiles (GDAL last-source-wins).
-BORDER_ZONES = {
-    "canada": (41.0, 55.0, -141.0, -52.0),
-    "mexico": (25.0, 33.0, -118.0, -96.0),
-}
-
-# Output nodata must match the pipeline convention (build_vrt.py srcNodata).
-OUTPUT_NODATA = -9999
-
-# Glob pattern for NHDPlus _fixed_ elevation tiles (written by compute_slope_aspect.py).
-NHDPLUS_FIXED_PATTERN = "NEDSnapshot_merged_fixed_*.tif"
-FILL_DIRS = {"copernicus_fill"}
-
-
-def _elapsed(t0: float) -> str:
-    secs = time.time() - t0
-    m, s = divmod(int(secs), 60)
-    return f"{m}m {s:02d}s" if m else f"{s}s"
-
-
-def _build_nhdplus_vrt(nhd_merged_dir: Path, output_vrt: Path) -> Path:
-    """Build a VRT from NHDPlus _fixed_ tiles only (no fill layers).
-
-    Returns the VRT path.  Raises FileNotFoundError if no NHDPlus _fixed_
-    tiles are found.  Raises RuntimeError if gdal.BuildVRT fails.
-    """
-    primary_files = sorted(
-        f for f in nhd_merged_dir.glob(f"*/{NHDPLUS_FIXED_PATTERN}")
-        if f.parent.name not in FILL_DIRS
-    )
-    if not primary_files:
-        raise FileNotFoundError(
-            f"No NHDPlus _fixed_ tiles found in {nhd_merged_dir}. "
-            "Run compute_slope_aspect.py first."
-        )
-    vrt_options = gdal.BuildVRTOptions(
-        resolution="highest", srcNodata=str(OUTPUT_NODATA),
-    )
-    vrt_ds = gdal.BuildVRT(
-        str(output_vrt), [str(f) for f in primary_files], options=vrt_options,
-    )
-    if vrt_ds is None:
-        raise RuntimeError("gdal.BuildVRT failed for NHDPlus-only VRT")
-    vrt_ds.FlushCache()
-    del vrt_ds
-    return output_vrt
-
-
-def _build_composite_vrt(
-    copernicus_elev: Path, nhdplus_vrt: Path, output_vrt: Path,
-) -> Path:
-    """Build a composite elevation VRT: Copernicus first (low priority),
-    NHDPlus last (high priority, wins in overlap).
-    """
-    vrt_options = gdal.BuildVRTOptions(
-        resolution="highest", srcNodata=str(OUTPUT_NODATA),
-    )
-    vrt_ds = gdal.BuildVRT(
-        str(output_vrt),
-        [str(copernicus_elev), str(nhdplus_vrt)],
-        options=vrt_options,
-    )
-    if vrt_ds is None:
-        raise RuntimeError("gdal.BuildVRT failed for composite elevation")
-    vrt_ds.FlushCache()
-    del vrt_ds
-    return output_vrt
-
-
-def _compute_fill_mask(
-    copernicus_elev: Path,
-    nhdplus_vrt: Path,
-    geotransform: tuple,
-    rows: int,
-    cols: int,
-) -> np.ndarray:
-    """Compute the fill-zone boolean mask.
-
-    Fill zone = pixels where Copernicus has valid data AND NHDPlus has nodata.
-    Both sources are warped to match the target grid defined by geotransform/rows/cols.
-
-    Returns a boolean ndarray of shape (rows, cols).
-    """
-    output_bounds = [
-        geotransform[0],
-        geotransform[3] + rows * geotransform[5],
-        geotransform[0] + cols * geotransform[1],
-        geotransform[3],
-    ]
-    warp_kwargs = dict(
-        format="MEM",
-        outputBounds=output_bounds,
-        xRes=abs(geotransform[1]),
-        yRes=abs(geotransform[5]),
-        dstNodata=OUTPUT_NODATA,
-        srcNodata=OUTPUT_NODATA,
-    )
-
-    # Read Copernicus elevation aligned to raw raster extent
-    cop_ds = gdal.Warp("", str(copernicus_elev), **warp_kwargs)
-    if cop_ds is None:
-        raise RuntimeError(
-            f"gdal.Warp to MEM failed for Copernicus readback: "
-            f"{copernicus_elev} — {gdal.GetLastErrorMsg()}"
-        )
-    cop_data = cop_ds.GetRasterBand(1).ReadAsArray().astype(np.float32)
-    del cop_ds
-
-    # Read NHDPlus VRT aligned to raw raster extent
-    nhd_ds = gdal.Warp("", str(nhdplus_vrt), **warp_kwargs)
-    if nhd_ds is None:
-        raise RuntimeError(
-            f"gdal.Warp to MEM failed for NHDPlus VRT readback: "
-            f"{nhdplus_vrt} — {gdal.GetLastErrorMsg()}"
-        )
-    nhd_data = nhd_ds.GetRasterBand(1).ReadAsArray().astype(np.float32)
-    del nhd_ds
-
-    # Shape guard — warp can produce off-by-one due to floating-point bounds
-    if cop_data.shape != (rows, cols) or nhd_data.shape != (rows, cols):
-        raise RuntimeError(
-            f"Shape mismatch after warp: expected ({rows}, {cols}), "
-            f"got cop={cop_data.shape}, nhd={nhd_data.shape}"
-        )
-
-    # Fill zone: Copernicus valid AND NHDPlus nodata.
-    # -9999 is exactly representable in float32, so equality comparison is safe.
-    return (cop_data != OUTPUT_NODATA) & (nhd_data == OUTPUT_NODATA)
-
-
-def _apply_fill_mask(
-    raw_raster: Path,
-    fill_mask: np.ndarray,
-    output: Path,
-) -> None:
-    """Apply a pre-computed fill mask to a raster, writing the masked result.
-
-    Pixels where fill_mask is True retain their raw value; all others become nodata.
-    """
-    raw_ds = gdal.Open(str(raw_raster))
-    if raw_ds is None:
-        raise RuntimeError(
-            f"gdal.Open failed for raw raster: {raw_raster} — {gdal.GetLastErrorMsg()}"
-        )
-    raw_data = raw_ds.GetRasterBand(1).ReadAsArray().astype(np.float32)
-    geotransform = raw_ds.GetGeoTransform()
-    projection = raw_ds.GetProjection()
-    rows, cols = raw_data.shape
-    del raw_ds
-
-    if fill_mask.shape != (rows, cols):
-        raise RuntimeError(
-            f"Shape mismatch in _apply_fill_mask for {output.name}: "
-            f"raw=({rows}, {cols}), mask={fill_mask.shape}"
-        )
-
-    masked = np.where(fill_mask, raw_data, np.float32(OUTPUT_NODATA))
-
-    # Write output
-    driver = gdal.GetDriverByName("GTiff")
-    if driver is None:
-        raise RuntimeError("GTiff driver not available — check GDAL installation")
-    out_ds = driver.Create(
-        str(output), cols, rows, 1, gdal.GDT_Float32,
-        options=[
-            "COMPRESS=LZW", "PREDICTOR=2", "TILED=YES",
-            "BLOCKXSIZE=512", "BLOCKYSIZE=512", "BIGTIFF=YES",
-        ],
-    )
-    if out_ds is None:
-        raise RuntimeError(
-            f"gdal driver.Create failed for output: {output} — {gdal.GetLastErrorMsg()}"
-        )
-    out_ds.SetGeoTransform(geotransform)
-    out_ds.SetProjection(projection)
-    out_band = out_ds.GetRasterBand(1)
-    out_band.SetNoDataValue(OUTPUT_NODATA)
-    err = out_band.WriteArray(masked)
-    if err != gdal.CE_None:
-        raise RuntimeError(
-            f"WriteArray failed for {output} — GDAL error code {err}: "
-            f"{gdal.GetLastErrorMsg()}"
-        )
-    out_ds.FlushCache()
-    del out_ds
+from gfv2_params.shared_rasters import SharedRastersContext
+from gfv2_params.shared_rasters.build_border_dem import build
 
 
 def main():
@@ -227,242 +27,19 @@ def main():
     args = parser.parse_args()
 
     logger = configure_logging("build_border_dem")
-    t_start = time.time()
 
     base = load_base_config(
         Path(args.base_config) if args.base_config else None,
         fabric=args.fabric,
     )
     data_root = Path(base["data_root"])
-
-    raw_dir = data_root / "input" / "copernicus_dem" / "raw"
-    nhd_merged_dir = data_root / "work" / "nhd_merged"
-    fill_dir = nhd_merged_dir / "copernicus_fill"
-    fill_dir.mkdir(parents=True, exist_ok=True)
-
-    elev_out = fill_dir / "NEDSnapshot_merged_fixed_copernicus.tif"
-    slope_out = fill_dir / "NEDSnapshot_merged_slope_copernicus.tif"
-    aspect_out = fill_dir / "NEDSnapshot_merged_aspect_copernicus.tif"
-
-    # --- Step 1: Compute tile list and download ---
-    logger.info("=== Step 1/5: Download Copernicus GLO-30 tiles ===")
-    all_labels = []
-    for zone_name, (south, north, west, east) in BORDER_ZONES.items():
-        labels = tiles_for_bbox(south, north, west, east)
-        logger.info("  %s zone: %d tiles (%.0f°N–%.0f°N, %.0f°W–%.0f°W)",
-                     zone_name, len(labels), south, north, abs(west), abs(east))
-        all_labels.extend(labels)
-
-    # Deduplicate (zones may overlap slightly)
-    all_labels = sorted(set(all_labels))
-    logger.info("  Total unique tiles: %d", len(all_labels))
-
-    t1 = time.time()
-    tile_paths = download_tiles(all_labels, raw_dir)
-    logger.info("  Download complete in %s: %d tiles available", _elapsed(t1), len(tile_paths))
-
-    if not tile_paths:
-        raise RuntimeError(
-            "No Copernicus tiles downloaded — check network access and tile labels"
-        )
-
-    # Warn on significant download shortfall
-    n_requested = len(all_labels)
-    n_downloaded = len(tile_paths)
-    if n_downloaded < n_requested:
-        shortfall_pct = 100 * (n_requested - n_downloaded) / n_requested
-        msg = (
-            f"Only {n_downloaded}/{n_requested} tiles downloaded "
-            f"({shortfall_pct:.0f}% shortfall) — border DEM may have coverage gaps"
-        )
-        if shortfall_pct > 20:
-            raise RuntimeError(msg)
-        logger.warning(msg)
-
-    # --- Step 2: Mosaic raw tiles and reproject ---
-    logger.info("=== Step 2/5: Mosaic → reproject to EPSG:5070 ===")
-    raw_vrt = fill_dir / "copernicus_raw.vrt"
-    vrt_ds = gdal.BuildVRT(
-        str(raw_vrt),
-        [str(p) for p in tile_paths],
+    ctx = SharedRastersContext(
+        data_root=data_root,
+        vpus=[],  # CONUS-once; discovers NHDPlus sources by glob
+        output_dir=data_root / "work",
+        force=args.force,
     )
-    if vrt_ds is None:
-        raise RuntimeError("gdal.BuildVRT failed for Copernicus raw tiles")
-    vrt_ds.FlushCache()
-    del vrt_ds
-    logger.info("  Raw VRT: %s (%d sources)", raw_vrt, len(tile_paths))
-
-    # Warp to EPSG:5070, 30m, nodata=-9999
-    if not elev_out.exists() or args.force:
-        logger.info("  Warping to EPSG:5070 at 30m (bilinear)...")
-        t2 = time.time()
-        warp_ds = gdal.Warp(
-            str(elev_out),
-            str(raw_vrt),
-            dstSRS="EPSG:5070",
-            xRes=30,
-            yRes=30,
-            resampleAlg="bilinear",
-            dstNodata=OUTPUT_NODATA,
-            outputType=gdal.GDT_Float32,
-            creationOptions=[
-                "COMPRESS=LZW",
-                "PREDICTOR=2",
-                "TILED=YES",
-                "BLOCKXSIZE=512",
-                "BLOCKYSIZE=512",
-                "BIGTIFF=YES",
-            ],
-        )
-        if warp_ds is None:
-            raise RuntimeError(
-                f"gdal.Warp failed: {raw_vrt} -> {elev_out} (EPSG:5070, 30m) "
-                f"— {gdal.GetLastErrorMsg()}"
-            )
-        warp_ds.FlushCache()
-        del warp_ds
-        logger.info("  Warp complete in %s: %s", _elapsed(t2), elev_out)
-    else:
-        logger.info("  Elevation fill already exists: %s", elev_out)
-
-    # --- Steps 3-5: Build composite, compute slope/aspect, mask ---
-    nhdplus_vrt = fill_dir / "nhdplus_only.vrt"
-    composite_vrt = fill_dir / "composite_elevation.vrt"
-    composite_clipped = fill_dir / "composite_elevation_clipped.tif"
-    slope_raw = fill_dir / "slope_raw.tif"
-    aspect_raw = fill_dir / "aspect_raw.tif"
-    intermediates = [slope_raw, aspect_raw, nhdplus_vrt, composite_vrt, composite_clipped]
-
-    if not slope_out.exists() or not aspect_out.exists() or args.force:
-        try:
-            # --- Step 3: Build composite elevation (Copernicus + NHDPlus) ---
-            logger.info("=== Step 3/5: Build composite elevation VRT ===")
-
-            _build_nhdplus_vrt(nhd_merged_dir, nhdplus_vrt)
-            logger.info("  NHDPlus-only VRT: %s", nhdplus_vrt)
-
-            _build_composite_vrt(elev_out, nhdplus_vrt, composite_vrt)
-            logger.info("  Composite VRT: %s", composite_vrt)
-
-            # Clip composite to Copernicus extent — the composite VRT covers the
-            # union of NHDPlus (all CONUS) + Copernicus, but we only need slope/aspect
-            # for the Copernicus extent. Loading the full union into RichDEM would be
-            # an unnecessary memory burden. The NHDPlus data in the overlap zone is
-            # still included (it falls within the Copernicus bounds at 41-55°N).
-
-            # Get Copernicus extent to clip the composite
-            cop_ds = gdal.Open(str(elev_out))
-            if cop_ds is None:
-                raise RuntimeError(
-                    f"gdal.Open failed for Copernicus elevation: {elev_out} "
-                    f"— {gdal.GetLastErrorMsg()}"
-                )
-            cop_gt = cop_ds.GetGeoTransform()
-            cop_cols = cop_ds.RasterXSize
-            cop_rows = cop_ds.RasterYSize
-            cop_bounds = [
-                cop_gt[0],
-                cop_gt[3] + cop_rows * cop_gt[5],
-                cop_gt[0] + cop_cols * cop_gt[1],
-                cop_gt[3],
-            ]
-            del cop_ds
-
-            logger.info("  Clipping composite to Copernicus extent...")
-            # No reprojection, same resolution — nearest-neighbour default is acceptable
-            clip_ds = gdal.Warp(
-                str(composite_clipped),
-                str(composite_vrt),
-                outputBounds=cop_bounds,
-                xRes=30, yRes=30,
-                dstNodata=OUTPUT_NODATA,
-                srcNodata=OUTPUT_NODATA,
-                outputType=gdal.GDT_Float32,
-                creationOptions=[
-                    "COMPRESS=LZW", "PREDICTOR=2", "TILED=YES",
-                    "BLOCKXSIZE=512", "BLOCKYSIZE=512", "BIGTIFF=YES",
-                ],
-            )
-            if clip_ds is None:
-                raise RuntimeError(
-                    f"gdal.Warp failed for composite clipping: {composite_vrt} -> "
-                    f"{composite_clipped} — {gdal.GetLastErrorMsg()}"
-                )
-            clip_ds.FlushCache()
-            del clip_ds
-
-            # --- Step 4: Compute slope/aspect from clipped composite ---
-            logger.info("=== Step 4/5: Compute slope/aspect from composite via RichDEM ===")
-            logger.info("  Loading clipped composite DEM: %s", composite_clipped)
-            t3 = time.time()
-            dem = rd.LoadGDAL(str(composite_clipped), no_data=OUTPUT_NODATA)
-
-            logger.info("  Computing slope (degrees)...")
-            slope = rd.TerrainAttribute(dem, attrib="slope_degrees")
-            rd.SaveGDAL(str(slope_raw), slope)
-            if not slope_raw.exists() or slope_raw.stat().st_size == 0:
-                raise RuntimeError(f"rd.SaveGDAL produced no output for slope: {slope_raw}")
-            logger.info("  Raw slope saved: %s", slope_raw)
-
-            logger.info("  Computing aspect...")
-            aspect = rd.TerrainAttribute(dem, attrib="aspect")
-            rd.SaveGDAL(str(aspect_raw), aspect)
-            if not aspect_raw.exists() or aspect_raw.stat().st_size == 0:
-                raise RuntimeError(f"rd.SaveGDAL produced no output for aspect: {aspect_raw}")
-            logger.info("  Raw aspect saved: %s", aspect_raw)
-            logger.info("  Slope/aspect computation complete in %s", _elapsed(t3))
-
-            # Free RichDEM arrays before masking to reduce peak memory
-            del dem, slope, aspect
-
-            # --- Step 5: Mask slope/aspect to fill zone ---
-            logger.info("=== Step 5/5: Mask slope/aspect to fill zone ===")
-            t4 = time.time()
-
-            # Compute fill mask once — reuse for both slope and aspect to halve
-            # I/O (avoids re-reading Copernicus and NHDPlus rasters).
-            ref_ds = gdal.Open(str(slope_raw))
-            if ref_ds is None:
-                raise RuntimeError(
-                    f"gdal.Open failed for slope raw raster: {slope_raw} "
-                    f"— {gdal.GetLastErrorMsg()}"
-                )
-            ref_gt = ref_ds.GetGeoTransform()
-            ref_rows = ref_ds.RasterYSize
-            ref_cols = ref_ds.RasterXSize
-            del ref_ds
-
-            fill_mask = _compute_fill_mask(
-                elev_out, nhdplus_vrt, ref_gt, ref_rows, ref_cols,
-            )
-
-            logger.info("  Masking slope to fill zone...")
-            _apply_fill_mask(slope_raw, fill_mask, slope_out)
-            logger.info("  Masked slope saved: %s", slope_out)
-
-            logger.info("  Masking aspect to fill zone...")
-            _apply_fill_mask(aspect_raw, fill_mask, aspect_out)
-            logger.info("  Masked aspect saved: %s", aspect_out)
-
-            logger.info("  Masking complete in %s", _elapsed(t4))
-
-        finally:
-            # Always clean up intermediates, even on failure
-            for f in intermediates:
-                if f.exists():
-                    f.unlink()
-                    logger.debug("  Cleaned up intermediate: %s", f)
-            logger.info("  Cleaned up intermediate files")
-    else:
-        logger.info("  Slope/aspect outputs already exist — skipping")
-
-    # Clean up raw VRT (created before the skip guard)
-    if raw_vrt.exists():
-        raw_vrt.unlink()
-
-    logger.info("=== build_border_dem complete in %s ===", _elapsed(t_start))
-    logger.info("  Outputs in: %s", fill_dir)
-    logger.info("  Run build_vrt.py to rebuild VRTs with the fill layer.")
+    build({}, ctx, logger)
 
 
 if __name__ == "__main__":
