+++ scripts/build_vpu_landmask.py
@@ -1,110 +1,20 @@
 """Build the per-VPU HRU-fabric land mask consumed by the TWI pipeline.
 
-Rasterises HRUs filtered to ``vpu == <vpu>`` onto the per-VPU Hydrodem grid
-(``Hydrodem_merged_<vpu>.tif``), producing
-``work/nhd_merged/<vpu>/land_mask_<vpu>.tif`` — a uint8 1/255 binary raster
-where 1 = inside one of this VPU's HRUs, 255 = outside.
+Thin CLI shell over the ``gfv2_params.shared_rasters.build_vpu_landmask``
+library entrypoint. Preserves the original argparse interface so existing
+sbatch jobs keep working unchanged.
 
-Why a per-VPU mask instead of the CONUS ``land_mask.tif`` from PR #69: the
-per-VPU Hydrodem is a roughly rectangular tile that bulges past its VPU's
-HRU boundary on inland flanks (the western boundary against adjacent VPUs,
-and the northern boundary against the international line). The CONUS
-``land_mask.tif`` covers every VPU's HRUs, so it reports "land" in those
-bulges wherever an adjacent VPU happens to drape coverage. That left the
-open-source TWI raster (``Twi_hydrodem_<vpu>.tif``) and the canonical
-``Twi_merged_<vpu>.tif`` with values along the priority-flood-to-nodata
-boundary — striped patterns inside the Hydrodem footprint but outside
-this VPU's HRU fabric.
-
-This mask constrains each per-VPU TWI product to just the HRUs that
-belong to that VPU. The CONUS depstor ``land_mask.tif`` is left alone —
-depstor's per-HRU zonal stats don't care about cross-VPU mask bleed.
-
-Fabric-independent by design: the open-source TWI rasters this mask
-constrains are global products (one per VPU under ``work/nhd_merged/``),
-so the masks are also global. The HRU source is the canonical CONUS
-fabric (``gfv2_nhru_merged.gpkg``); per-fabric HRU subsets are subsets
-of that gpkg and filtering by the ``vpu`` column is sufficient. The
-script takes no ``--fabric`` argument for this reason.
-
-Output: ``{data_root}/work/nhd_merged/<vpu>/land_mask_<vpu>.tif``
+For unified DAG-style invocation across all shared rasters, prefer:
+    python scripts/build_shared_rasters.py --config configs/shared_rasters.yml
 """
 
 import argparse
-import time
 from pathlib import Path
 
-import geopandas as gpd
-
-from gfv2_params.config import VPU_RASTER_MAP, load_config, require_config_key
-from gfv2_params.depstor import RasterInfo, rasterize_binary, write_uint8_binary
+from gfv2_params.config import load_config, require_config_key
 from gfv2_params.log import configure_logging
-
-
-def _elapsed(t0: float) -> str:
-    secs = time.time() - t0
-    m, s = divmod(int(secs), 60)
-    return f"{m}m {s:02d}s" if m else f"{s}s"
-
-
-def _load_hru(path: Path, layer: str, logger):
-    try:
-        return gpd.read_file(path, layer=layer, use_arrow=True)
-    except ImportError:
-        logger.warning("PyArrow unavailable for vector load; falling back to fiona.")
-        return gpd.read_file(path, layer=layer)
-
-
-def build_vpu_landmask(
-    template_path: Path,
-    hru_gpkg: Path,
-    hru_layer: str,
-    vpu: str,
-    output_path: Path,
-    logger,
-    vpu_column: str = "vpu",
-):
-    """Rasterise HRUs filtered to this VPU onto the per-VPU template grid.
-
-    Pulled out of ``main`` so unit tests can exercise the build without going
-    through argument parsing and config-file resolution. Comparison is
-    string-equality after zero-padding to two characters, so values stored
-    as int (``1``) or zero-padded string (``"01"``) both match.
-    """
-    info = RasterInfo.from_path(template_path)
-    hru_gdf = _load_hru(hru_gpkg, hru_layer, logger)
-
-    if vpu_column not in hru_gdf.columns:
-        raise KeyError(
-            f"HRU gpkg {hru_gpkg} layer '{hru_layer}' has no '{vpu_column}' "
-            f"column; cannot filter by VPU. Available columns: {sorted(hru_gdf.columns)}"
-        )
-
-    # Per-VPU rasters use simple two-character codes (`03`, `10`), but the HRU
-    # fabric stores sub-region codes (`03N`/`03S`/`03W`, `10L`/`10U`) — there
-    # is no plain `03` or `10` row in gfv2_nhru_merged.gpkg. VPU_RASTER_MAP
-    # encodes the sub-region -> raster-VPU mapping, so we invert it here to
-    # build the set of acceptable HRU vpu values for this raster VPU.
-    vpu_key = str(vpu).zfill(2)
-    sub_vpus = {sub for sub, parent in VPU_RASTER_MAP.items() if parent == vpu_key}
-    acceptable_vpus = {vpu_key} | sub_vpus
-    logger.info("  Filtering HRUs by %s in %s", vpu_column, sorted(acceptable_vpus))
-
-    matches = hru_gdf[vpu_column].astype(str).isin(acceptable_vpus)
-    hru_gdf = hru_gdf[matches]
-    hru_gdf = hru_gdf[hru_gdf.geometry.notna() & ~hru_gdf.geometry.is_empty]
-    if hru_gdf.empty:
-        raise ValueError(
-            f"No HRUs matched {vpu_column} in {sorted(acceptable_vpus)} in "
-            f"{hru_gpkg}:{hru_layer}. Verify the VPU code and the fabric gpkg."
-        )
-
-    # all_touched=True for the same reason as build_depstor_landmask: stay
-    # inclusive at thin HRU edges; create_zonal_params remains the precise
-    # arbiter downstream.
-    binary = rasterize_binary(hru_gdf, info, all_touched=True)
-    write_uint8_binary(binary, info, output_path)
-    return binary, len(hru_gdf)
+from gfv2_params.shared_rasters import SharedRastersContext
+from gfv2_params.shared_rasters.build_vpu_landmask import build, build_vpu_landmask  # noqa: F401 — re-exported for tests
 
 
 def main():
@@ -116,53 +26,39 @@ def main():
     args = parser.parse_args()
 
     logger = configure_logging("build_vpu_landmask")
-    t_start = time.time()
 
     # No --fabric: the per-VPU land mask is a fabric-independent product. The
-    # config pins the HRU source to the canonical CONUS `gfv2_nhru_merged.gpkg`
-    # and the output to work/nhd_merged/<vpu>/ — both fabric-free locations.
-    # `load_config` still goes through fabric resolution to set `{data_root}`,
-    # but the step-config's `hru_gpkg` / `hru_layer` win over any fabric-profile
-    # defaults via the standard step-overrides-base merge.
+    # config pins the HRU source to the canonical CONUS gfv2_nhru_merged.gpkg
+    # and the output to work/nhd_merged/<vpu>/.
     config = load_config(
         Path(args.config),
         vpu=args.vpu,
         base_config_path=Path(args.base_config) if args.base_config else None,
     )
 
-    template_path = Path(require_config_key(config, "template_raster", "build_vpu_landmask"))
-    hru_gpkg = Path(require_config_key(config, "hru_gpkg", "build_vpu_landmask"))
+    template_raster = require_config_key(config, "template_raster", "build_vpu_landmask")
+    hru_gpkg = require_config_key(config, "hru_gpkg", "build_vpu_landmask")
     hru_layer = require_config_key(config, "hru_layer", "build_vpu_landmask")
-    output_path = Path(config["output_raster"])
+    output_raster = config["output_raster"]
 
-    if not template_path.exists():
-        raise FileNotFoundError(f"Per-VPU Hydrodem template not found: {template_path}")
-    if not hru_gpkg.exists():
-        raise FileNotFoundError(f"HRU fabric gpkg not found: {hru_gpkg}")
-
-    logger.info("=== build_vpu_landmask (VPU %s) ===", args.vpu)
-    logger.info("Template  : %s", template_path)
-    logger.info("HRU fabric: %s (layer=%s, filtered to vpu=%s)", hru_gpkg, hru_layer, args.vpu)
-    logger.info("Output    : %s", output_path)
-
-    if output_path.exists() and not args.force:
-        logger.info("Output already exists — skipping (pass --force to rebuild)")
-        return
-
-    info = RasterInfo.from_path(template_path)
-    logger.info("Template grid: %dx%d, CRS=%s", info.width, info.height, info.crs)
-
-    t1 = time.time()
-    binary, n_polys = build_vpu_landmask(
-        template_path, hru_gpkg, hru_layer, args.vpu, output_path, logger,
+    data_root = Path(config["data_root"])
+    ctx = SharedRastersContext(
+        data_root=data_root,
+        vpus=[args.vpu],
+        output_dir=data_root / "work",
+        force=args.force,
     )
-    n_land = int((binary == 1).sum())
-    logger.info(
-        "  Rasterised %d HRU polygons in %s | %d land cells (%.2f%% of grid)",
-        n_polys, _elapsed(t1), n_land, 100 * n_land / binary.size,
-    )
-
-    logger.info("=== build_vpu_landmask complete in %s ===", _elapsed(t_start))
+    # CLI-resolved paths already have {vpu} substituted (via load_config). The
+    # builder's per-VPU loop calls .replace("{vpu}", vpu) which is a no-op on
+    # strings with no placeholder, so passing the resolved paths through works
+    # for both single-VPU (CLI) and multi-VPU (orchestrator) invocations.
+    step_cfg = {
+        "template_raster": str(template_raster),
+        "hru_gpkg": hru_gpkg,
+        "hru_layer": hru_layer,
+        "output_raster": str(output_raster),
+    }
+    build(step_cfg, ctx, logger)
 
 
 if __name__ == "__main__":
